--- cj10-screaming-snakecases/Engine/ImageProcesser.py
@@ -1,18 +1,0 @@
-# This file should contain image processing functions, as ImageTools contain image manipulation functions
-# Potential things to add (@Envsxion)
-
-# Forgery detection: Implement algorithms to detect image tampering or forgery.
-# Steganography tools: Kinda self explanatory.
-# Hashing and authentication: Implement image hashing and authentication mechanisms.
-# Noise Analysis
-# Blur Detection
-# Content-Based Image Retrieval (CBIR): dunno how but ig we can just use an api or send it off to google images
-# OCR (Optical Character Recognition)
-# Geolocation analysis
-# Image Comparison (finding mathematical differences between images to reveal the flag??)
-# File Format Analysis (images inside an image are a great way to f ppl over)
-
-
-class ImageProcessor:
-    def __init__(self) -> None:
-        pass
