--- cj10-screaming-snakecases/Engine/ImageViewer.py
@@ -1,9 +1,0 @@
-# This file contains the class ImageViewer, which should show pixel-level statistics of a certain region of interest
-# For example: Show histograms, color channels, and pixel-level details. (@Envxsion)
-
-# This will have to be programmed in conjunction with the GUI soon.
-
-
-class ImageViewer():
-    def __init__(self) -> None:
-        pass
