+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "pixelheist"
+name = "Name"
 version = "0.1.0"
 description = ""
 authors = ["Author 1 <email@mail.com>"]
@@ -13,20 +13,18 @@ pyside6 = "^6.5.2"
 
 # Engine
 Pillow = "^10.0.0"
-geopy = "^2.4.0"
-numpy = "^1.25.2"
+flake8 = "^6.1.0"
+isort = "^5.12.0"
 
 [tool.poetry.dev-dependencies]
 # Base tools
 flake8 = "~6.1.0"
 isort = "~5.12.0"
-pre-commit = "~3.4.0"
+pre-commit = "~3.3.3"
 
 # Flake8 plugins, see https://github.com/python-discord/code-jam-template/tree/main#plugin-list
 flake8-docstrings = "~1.7.0"
 
-# For testing the engine
-pygame = "^2.5.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
